feat(controllers): Adds PollsController feat(serializers): Adds PollSerializer & PollStatsSerializer feat(services): Adds PollOptionsService
PollsController handles all endpoints needed by the app, that would respectively be routed to its actions as:
- get /poll/:id to #show
- get /poll/:id/stats to #stats
- post /poll to #create
- post /poll/:id/vote to #vote

Adds two serializers to format response for each GET request as documented.

Adds PollOptionsService to handle the conversion of the documented POST payload (which serves PollOptions as an array of strings, representing their descriptions) converting
it into an hash of PollOption objects.
This behaviour could be avoided if payload was served in a more rails-ish way (nested_attributes), but preserving the test specifications was treated as high priority.