feat(Docker): Dockerizes the application
Sets up docker-compose.yml and docker-compose.override.yml aswell as docker/app/dev.app.Dockerfile to act as a development container environment.
On this setup a database (db) service is supplied to fill the needs of a local postgresql database when running the app.

Also sets up files for establishing a production environment serving the application through the web:
- docker-compose.production.yml
- docker-compose.production.override.yml
- docker/app/prod.app.Dockerfile
- docker/web/web.Dockerfile
- docker/web/nginx.conf
- .env
- .env_production

This docker-compose set up relies on serving the application through nginx upon the web service, the application itself through app service and it's also expected that
credentials for a remote database connection be supplied into .env_production file when deploying the production version.